+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 5.1.0-beta.6
 
 * Fix: set min sdk version of sentry-android-fragment to API 14 (#1608)
 * Fix: Ser/Deser of the UserFeedback from cached envelope (#1611)